+++ crypto/src/crypto/digests/RipeMD128Digest.cs
@@ -116,10 +116,10 @@ namespace Org.BouncyCastle.Crypto.Digests
         {
             base.Reset();
 
-            H0 = unchecked((int) 0x67452301);
+            H0 = unchecked(0x67452301);
             H1 = unchecked((int) 0xefcdab89);
             H2 = unchecked((int) 0x98badcfe);
-            H3 = unchecked((int) 0x10325476);
+            H3 = unchecked(0x10325476);
 
             xOff = 0;
 
@@ -206,7 +206,7 @@ namespace Org.BouncyCastle.Crypto.Digests
             int x,
             int s)
         {
-            return RL(a + F2(b, c, d) + x + unchecked((int) 0x5a827999), s);
+            return RL(a + F2(b, c, d) + x + unchecked(0x5a827999), s);
         }
 
         private int F3(
@@ -217,7 +217,7 @@ namespace Org.BouncyCastle.Crypto.Digests
             int x,
             int s)
         {
-            return RL(a + F3(b, c, d) + x + unchecked((int) 0x6ed9eba1), s);
+            return RL(a + F3(b, c, d) + x + unchecked(0x6ed9eba1), s);
         }
 
         private int F4(
@@ -250,7 +250,7 @@ namespace Org.BouncyCastle.Crypto.Digests
             int x,
             int s)
         {
-        return RL(a + F2(b, c, d) + x + unchecked((int) 0x6d703ef3), s);
+        return RL(a + F2(b, c, d) + x + unchecked(0x6d703ef3), s);
         }
 
         private int FF3(
@@ -261,7 +261,7 @@ namespace Org.BouncyCastle.Crypto.Digests
             int x,
             int s)
         {
-        return RL(a + F3(b, c, d) + x + unchecked((int) 0x5c4dd124), s);
+        return RL(a + F3(b, c, d) + x + unchecked(0x5c4dd124), s);
         }
 
         private int FF4(
@@ -272,7 +272,7 @@ namespace Org.BouncyCastle.Crypto.Digests
             int x,
             int s)
         {
-        return RL(a + F4(b, c, d) + x + unchecked((int) 0x50a28be6), s);
+        return RL(a + F4(b, c, d) + x + unchecked(0x50a28be6), s);
         }
 
         internal override void ProcessBlock()
